drf_api/requirements.txt
# django 
djangorestframework==3.15.2
drf-spectacular==0.28.0
django-environ==0.11.2
django-filter==24.3
django-cors-headers==4.6.0
django-extensions
# databases
psycopg2-binary==2.9.10
pandas==2.2.3
influxdb-client==1.48.0

# tests
pytest==8.3.3
pytest-django==4.9.0
pytest-factoryboy==2.7.0

# utils
factory_boy==3.2.1


